requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
requests==2.32.3
python-dotenv==1.0.1